Update Kernel & Compiler:
Kernel.newspeak-eliot.671
Fix implementation of MethodContext>>numTemps.

Kernel.newspeak-eliot.670
Move the code to set the owning method of an
AdditionalMethodState to the setter away from clients.
Add an anti-sharing check.

Refactor unusedBytecode error in InstructionStream
to allow sending of unusedBytecode to client if it
understands the message.

Compiler.newspeak-eliot.227
Lift the number of literals limit in EncoderForNewsqueakV4.