add version of #changeRecordForOverridenMethod: that does not use so many file handles
When building our nsboot images, we'll run out of file handles from time to time.
This was traced down to #changeRecordForOverriddenMethod: and it copying the
system sources for each method in a package.  With this change, a cached
file handle is used instead.  For now, postpatches seemed like the best place
to put this.  An alternative is to create a new PackageInfo version (or see
if the squeak trunk has this or similar fix and use that version).