Update Network packaged used in the system (and more, read below)
This updates the Network package to a newer version. It includes
the modified version of UUID class>>fromString: that Newspeak
needs for coping with some old Monticello package versions that
have UUIDs with @ in them.

Also included are some tweaks to ensure that the Monticello
package code is using CurrentReadOnlySourceFiles when snapshotting
methods.  This prevents a copy of the source files to be made
for each snapshotted method.  This is vital for us at Cadence
because we build on machines where the file limit can not
build increased and we run into this limit quite often when building.