Add and boot Kernel.newspeak-eliot.676:
*TEMPORARILY* restore a simpified version of the hack "fix"
to ContextPart>>runUntilErrorOrReturnFrom: that apparently
rescues the Hopscotch Newspeak evaluator.

While the evaluator is broken without this, this patch breaks teh Squeak and Hopscoth debuggers with non-local return.

Try this:
	self halt.
	(1 to: 5) do: [:i| i = 1 ifTrue: [^#ok]. #bad]

Without this patch, stepping over the value: send in
Interval>>do: does what one expects; returns to the
sender of the doit.
With the patch, the debugger continues to evaluate the
block with 2 to: 5.

So caveat emptor!!